+++ Source/GateRimSG1/Jaffa/HediffCompProperties_TretoninSubstitution.cs
@@ -1,0 +1,12 @@
+using Verse;
+
+namespace GateRimSG1.Jaffa
+{
+    public class HediffCompProperties_TretoninSubstitution : HediffCompProperties
+    {
+        public HediffCompProperties_TretoninSubstitution()
+        {
+            compClass = typeof(HediffComp_TretoninSubstitution);
+        }
+    }
+}